+++ AmazonChimeSDK/AmazonChimeSDKTests/audiovideo/metric/ObservableMetricTests.swift
@@ -28,4 +28,21 @@ class ObservableMetricTests: XCTestCase {
         XCTAssertEqual(ObservableMetric.contentShareVideoSendFps.description, "contentShareVideoSendFps")
         XCTAssertEqual(ObservableMetric.contentShareVideoSendRttMs.description, "contentShareVideoSendRttMs")
     }
+
+    func testIsContentShareMetric() {
+        XCTAssertFalse(ObservableMetric.audioReceivePacketLossPercent.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.audioSendPacketLossPercent.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoAvailableSendBandwidth.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoAvailableReceiveBandwidth.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoSendBitrate.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoSendPacketLossPercent.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoSendFps.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoSendRttMs.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoReceiveBitrate.isContentShareMetric)
+        XCTAssertFalse(ObservableMetric.videoReceivePacketLossPercent.isContentShareMetric)
+        XCTAssertTrue(ObservableMetric.contentShareVideoSendBitrate.isContentShareMetric)
+        XCTAssertTrue(ObservableMetric.contentShareVideoSendPacketLossPercent.isContentShareMetric)
+        XCTAssertTrue(ObservableMetric.contentShareVideoSendFps.isContentShareMetric)
+        XCTAssertTrue(ObservableMetric.contentShareVideoSendRttMs.isContentShareMetric)
+    }
 }